requirements.txt
python-telegram-bot[job-queue]==22.5
robokassa
python-dotenv
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pytz
fastapi
uvicorn[standard]
python-multipart

